@@ -1,5 +1,4 @@
-using System.Data.SqlClient;
-using Microsoft.Extensions.Logging;
+using Microsoft.Data.SqlClient;
 using SqlKata.Compilers;
 using SqlKata.Execution;
 
