@@ -28,7 +28,7 @@ public class QueryApiTest
     }
 
     [Fact]
-    public async void TestRelatedData()
+    public async Task TestRelatedData()
     {
         var postId =await PrepareOneRelatedData();
         var query = GetPostQuery();
@@ -42,7 +42,7 @@ public class QueryApiTest
 
 
     [Fact]
-    public async void List()
+    public async Task List()
     {
         await PrepareSimpleData();
         var query = GetPostQuery();
@@ -54,7 +54,7 @@ public class QueryApiTest
     }
 
     [Fact]
-    public async void Many()
+    public async Task Many()
     {
         await PrepareSimpleData();
         var query = GetPostQuery();
